+++ test/integration/toolchains/cmake/cmakedeps2/test_cmakedeps.py
@@ -1,14 +1,10 @@
-import os
 import re
 import textwrap
-import platform
 
 import pytest
-from conan.test.utils.mocks import ConanFileMock
 
 from conan.test.assets.genconanfile import GenConanfile
 from conan.test.utils.tools import TestClient
-from conan.tools.files import replace_in_file
 
 new_value = "will_break_next"
 
@@ -213,134 +209,6 @@ def test_cmakeconfigdeps_recipe():
     assert "WARN: Using the new CMakeConfigDeps generator" in c.out
     c.run("install app -c tools.cmake.cmakedeps:new=recipe_will_break")
     assert "WARN: Using the new CMakeConfigDeps generator" in c.out
-
-
-@pytest.mark.skipif(platform.system() != "Linux", reason="Only linux")
-def test_cpp_info_sources():
-    c = TestClient()
-    c.run("new cmake_lib -d name=hello -d version=1.0")
-
-    cmakelists = textwrap.dedent("""
-        cmake_minimum_required(VERSION 3.15)
-        project(hello CXX)
-
-        add_library(hello INTERFACE)
-        target_include_directories(hello INTERFACE include)
-        set_target_properties(hello PROPERTIES PUBLIC_HEADER "include/hello.h")
-        target_sources(hello INTERFACE
-                       "include/hello.h"
-                       "src/hello.cpp")
-
-        install(TARGETS hello)
-        install(FILES src/hello.cpp DESTINATION src)
-    """)
-    conanfile = textwrap.dedent("""
-        import os
-        from conan.tools.cmake import CMake, CMakeToolchain
-        from conan.tools.files import copy
-        from conan import ConanFile
-
-        class HelloConan(ConanFile):
-            name = "hello"
-            version = "1.0"
-            settings = "os", "arch", "compiler", "build_type"
-            exports_sources = "CMakeLists.txt", "src/*", "include/*"
-
-            def generate(self):
-                tc = CMakeToolchain(self)
-                tc.generate()
-
-            def package(self):
-                cmake = CMake(self)
-                cmake.install()
-
-            def build(self):
-                cmake = CMake(self)
-                cmake.configure()
-
-            def package_info(self):
-                self.cpp_info.sources = ["include/hello.h", "src/hello.cpp"]
-    """)
-    c.save({
-        "CMakeLists.txt": cmakelists,
-        "conanfile.py": conanfile
-    })
-    c.run(f"create . -c tools.cmake.cmakedeps:new={new_value}")
-    c.run(f"install --requires=hello/1.0 -g=CMakeConfigDeps "
-          f"-c tools.cmake.cmakedeps:new={new_value}")
-    cmake = c.load("hello-Targets-release.cmake")
-    assert "add_library(hello::hello INTERFACE IMPORTED)" in cmake
-    assert "set_property(TARGET hello::hello APPEND PROPERTY INTERFACE_SOURCES\n"\
-           "             $<$<CONFIG:RELEASE>:${hello_PACKAGE_FOLDER_RELEASE}/include/hello.h;" \
-           "${hello_PACKAGE_FOLDER_RELEASE}/src/hello.cpp>)" in cmake
-
-
-@pytest.mark.skipif(platform.system() != "Linux", reason="Only linux")
-def test_cpp_info_sources_with_component():
-    c = TestClient()
-    c.run("new cmake_lib -d name=hello -d version=1.0")
-
-    cmakelists = textwrap.dedent("""
-        cmake_minimum_required(VERSION 3.15)
-        project(hello CXX)
-
-        add_library(hello INTERFACE)
-        target_include_directories(hello INTERFACE include)
-        set_target_properties(hello PROPERTIES PUBLIC_HEADER "include/hello.h")
-        target_sources(hello INTERFACE
-                       "include/hello.h"
-                       "src/hello.cpp")
-
-        install(TARGETS hello)
-        install(FILES src/hello.cpp DESTINATION src)
-    """)
-    conanfile = textwrap.dedent("""
-        import os
-        from conan.tools.cmake import CMake, CMakeToolchain
-        from conan.tools.files import copy
-        from conan import ConanFile
-
-        class HelloConan(ConanFile):
-            name = "hello"
-            version = "1.0"
-            settings = "os", "arch", "compiler", "build_type"
-            exports_sources = "CMakeLists.txt", "src/*", "include/*"
-
-            def generate(self):
-                tc = CMakeToolchain(self)
-                tc.generate()
-
-            def package(self):
-                cmake = CMake(self)
-                cmake.install()
-
-            def build(self):
-                cmake = CMake(self)
-                cmake.configure()
-
-            def package_info(self):
-                self.cpp_info.components["my_comp"].sources = ["include/hello.h", "src/hello.cpp"]
-    """)
-    c.save({
-        "CMakeLists.txt": cmakelists,
-        "conanfile.py": conanfile
-    })
-
-    # Make test_package link with component's target
-    test_package_cmakelists_path = os.path.join(c.current_folder, "test_package", "CMakeLists.txt")
-    replace_in_file(ConanFileMock(), test_package_cmakelists_path, "hello::hello", "hello::my_comp")
-    test_package_cmakelists_content = c.load(test_package_cmakelists_path)
-    assert "target_link_libraries(example hello::my_comp)" in test_package_cmakelists_content
-
-    c.run(f"create . -c tools.cmake.cmakedeps:new={new_value}")
-    c.run(f"install --requires=hello/1.0 -g=CMakeConfigDeps "
-          f"-c tools.cmake.cmakedeps:new={new_value}")
-    cmake = c.load("hello-Targets-release.cmake")
-    assert "add_library(hello::hello INTERFACE IMPORTED)" in cmake
-    assert "add_library(hello::my_comp INTERFACE IMPORTED)" in cmake
-    assert "set_property(TARGET hello::my_comp APPEND PROPERTY INTERFACE_SOURCES\n"\
-           "             $<$<CONFIG:RELEASE>:${hello_PACKAGE_FOLDER_RELEASE}/include/hello.h;" \
-           "${hello_PACKAGE_FOLDER_RELEASE}/src/hello.cpp>)" in cmake
 
 
 def test_system_wrappers():
